Hotfix/55ace fix

See merge request jolly88-complex/frontend!7150

+++ ind-static-h5/src/boot/utils.js
@@ -109,7 +109,7 @@ export const convertToCommaAmount = (amount, isForceDecimal) => {
   if (isNonNumericString(amount)) {
     return amount;
   }
-  return parseInt(amount).toLocaleString("en-US", { minimumFractionDigits: isForceDecimal ? 2 : 0 });
+  return parseFloat(amount).toLocaleString("en-US", { minimumFractionDigits: isForceDecimal ? 2 : 0 });
 };
 
 function isNonNumericString(value) {
